refactor(tests): pull out field compilation tests
I've pulled out the tests about fields into their own test files. Previously they were in the struct
tests, which doesn't make much sense now that we've got unions, which also have fields.

Part of #12
